update OpenVHook to latest GTAV, add missing signatures for previous game versions

+++ OpenVHook/Input/InputHook.h
@@ -5,7 +5,7 @@
 
 typedef struct KeyState
 {
-	int lastUpTime;
+	ULONGLONG lastUpTime;
 	BOOL isWithAlt;
 	BOOL wasDownBefore;
 	BOOL isUpNow;
